+++ lua.go
@@ -48,6 +48,10 @@ func luaInit() {
 		}
 	})
 
+	lr.rl.RawInputCallback = func(r []rune) {
+		hooks.Em.Emit("hilbish.rawInput", string(r))
+	}
+
 	// Add more paths that Lua can require from
 	err := util.DoString(l, "package.path = package.path .. " + requirePaths)
 	if err != nil {